requirements.txt
numpy
pyyaml
torch
tensorboard
